resources/aws-init/requirements.txt
boto3
pycryptodome
requests
